[ALPS05568686] ccci: change data type
fix build error: shift count >= width of type.
add cast type for address value.

MTK-Commit-Id: 4fa2e4df4354af5218f4a4c8b060ba9b9331c3f5

CR-Id: ALPS05568686
Feature: Modem Interface Driver

+++ drivers/misc/mediatek/eccci/ccci_modem.c
@@ -1962,7 +1962,7 @@ static void ccci_md_mem_inf_prepare(int md_id,
 		tbl[add_num].md_view_phy = 0;
 		tbl[add_num].size = ro_rw_size;
 		tbl[add_num].ap_view_phy_lo32 = (u32)ro_rw_base;
-		tbl[add_num].ap_view_phy_hi32 = (u32)(ro_rw_base >> 32);
+		tbl[add_num].ap_view_phy_hi32 = (u32)((unsigned long long)ro_rw_base >> 32);
 		add_num++;
 	} else
 		CCCI_REPEAT_LOG(md_id, TAG, "%s add bank0/1 fail(%d)\n",
@@ -1972,7 +1972,7 @@ static void ccci_md_mem_inf_prepare(int md_id,
 		tbl[add_num].md_view_phy = 0x40000000;
 		tbl[add_num].size = ncrw_size;
 		tbl[add_num].ap_view_phy_lo32 = (u32)ncrw_base;
-		tbl[add_num].ap_view_phy_hi32 = (u32)(ncrw_base >> 32);
+		tbl[add_num].ap_view_phy_hi32 = (u32)((unsigned long long)ncrw_base >> 32);
 		add_num++;
 	} else
 		CCCI_REPEAT_LOG(md_id, TAG, "%s add bank4 nc fail(%d)\n",
@@ -1983,7 +1983,7 @@ static void ccci_md_mem_inf_prepare(int md_id,
 				get_md_smem_cachable_offset(md_id);
 		tbl[add_num].size = crw_size;
 		tbl[add_num].ap_view_phy_lo32 = (u32)crw_base;
-		tbl[add_num].ap_view_phy_hi32 = (u32)(crw_base >> 32);
+		tbl[add_num].ap_view_phy_hi32 = (u32)((unsigned long long)crw_base >> 32);
 		add_num++;
 	} else
 		CCCI_REPEAT_LOG(md_id, TAG, "%s add bank4 c fail(%d)\n",
